Add files via upload
This is a HellCoin. In the near future, we pray and hope to Jesus and Lucifer in hopes of being able to exchange these coins to you in a swap for a bit of silver, plastic or green. We only want you to pay for your sins now in advance. so you can pay for your sins before you go there. All payments induce cosmic karma clearance black-Friday style and real forgiveness for that sin, we are not involved with the law, only universal forgiveness obtainable not earthly dues. They have to be paid back also, and maybe they can forgive you also. 
 ,-,-.
/ (_o \
\ o ) /
 `-'-'